requirements.txt
djangorestframework
markdown
django-filter
drf-yasg